requirements-dev.txt
black>=26.1.0
isort
flake8
pytest
pytest-xdist
pytest-playwright
